% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{extractNodeData}
\alias{extractNodeData}
\title{Extracts node data into a data frame}
\usage{
extractNodeData(.tree, .index, .node = NULL)
}
\arguments{
\item{.tree}{- an "fTree" object}

\item{.index}{- index of the tree you wish to process}
}
