% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getNodeIndices}
\alias{getNodeIndices}
\title{Returns the indices of data points contained within a specific node}
\usage{
getNodeIndices(.tree, .node)
}
\arguments{
\item{.tree}{- An fTree object}

\item{.node}{- The code of the node you wish to extract (i.e. \code{.node="0lr"})}
}
\details{
The \code{.node} parameter should specify a path through the tree to
         the node you wish to extract for example: root-left-right-left = \code{.node='0lrl'}.
}
\author{
Ognjen Grujic \email{ogyg@stanford.edu}
}
