% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{matToList}
\alias{matToList}
\alias{matToList}
\title{A helper function for parallel computing (not exported)}
\usage{
matToList(.matrix, .chunksize)

matToList(.matrix, .chunksize)
}
\arguments{
\item{.matrix}{- A matrix to split into list;}

\item{.chunksize}{- A number of rows in each list element;}

\item{.matrix}{- A matrix you wish to divide into a list}

\item{.chunksize}{- Number of rows of matrix .matrix to be contained in each
element of the resulting list.}
}
\details{
This code simply splits rows of a matrix into a list where each element of the list
         contains .chunksize number of rows of the input matrix. If .chunksize > nrow(.matrix) it
         returns a list with one element, the .matrix itself.

This function simply splits provided matrix into N elements each containing
         .chunksize of rows of the original matrix. List elements are mutually exclusive
}
\author{
Ognjen Grujic (\email{ognjengr@gmail.com})

Ogy Grujic (\email{ogyg@stanford.edu})
}
