% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plotFtree}
\alias{plotFtree}
\title{Plots a (functional) regression tree}
\usage{
plotFtree(.object, .index = 1, .horizontal = FALSE, .lineWidth = 1,
  .depth = Inf, .labSize = 1, .labAng = 0, .ggReturn = FALSE,
  .ylimLow = 0, .ylimHigh = NA, .node = NULL, .round = NULL)
}
\arguments{
\item{.object}{- An \code{fTree} object or \code{.segments} object produced with \code{extractNodeData}}

\item{.index}{- Index of the tree you wish you plot (in the case of bootstrapped trees)}

\item{.horizontal}{- Whether to plot the tree with horizontal orientation}

\item{.lineWidth}{- The width of the tree lines}

\item{.depth}{- Maximum plotting depth (useful for large trees)}

\item{.labSize}{- The size of the node labels}

\item{.labAng}{- The angle of node labels}

\item{.ggReturn}{- Whether to return a \code{ggplot} object or not.}

\item{.ylimLow}{- Lower limit of y axis}

\item{.ylimHigh}{- Higher limit of y axis}

\item{.node}{- A node you wish to highlight. The node is specified as a path throught the tree from the root node.
Example: \code{.node = '0llr'} = 0 - root, l - left, l - left, r - right.
The left side branches of the node will be painted in red, while the right side will be painted in blue.}

\item{.round}{- The number of rounding digits for labels.}
}
\details{
This function relies on the \code{extractNodeData} function and \code{ggplot} plotting package.
         The function outputs publication ready images of the passed tree.
}
\author{
Ognjen Grujic (\email{ogyg@stanford.edu})
}
