% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plotNodeData}
\alias{plotNodeData}
\title{Plots node data of a functional regression tree}
\usage{
plotNodeData(ftreeObj, .treeIndex = 1, .node = "0", .ggRet = FALSE)
}
\arguments{
\item{ftreeObj}{- An fTree object}

\item{.treeIndex}{- Index of the tree you wish you plot}

\item{.node}{- A code for specific node to plot}

\item{.ggRet}{- Whether to return ggplot structure}
}
\details{
Depending on the type of the fitted tree this function will either plot raw functions
         or low dimensional representation with fPCA or MDS.
}
\author{
Ognjen Grujic \email{ogyg@stanford.edu}
}
