% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{returnLeafData}
\alias{returnLeafData}
\title{Returns leaf data, ID's and associated indices}
\usage{
returnLeafData(.ftree)
}
\arguments{
\item{.ftree}{- an Ftree data structure}
}
\description{
Returns leaf data, ID's and associated indices
}
\details{
It returns a data frame with two columns. The first column contains indices of the data relative to the original ordering, while the second column represents a unique leaf ID.
}
\author{
Ogy Grujic (\email{ognjengr@gmail.com})
}
