% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{varSensitivity}
\alias{varSensitivity}
\title{Produces a sensitivity/importance plot}
\usage{
varSensitivity(ftreeObj = NULL, mode = "plot")
}
\arguments{
\item{ftreeObj}{- An object of class "ftree"}

\item{mode}{- Mode you wish to run (plot - just plots, ggRet - returns a ggplot structure, data - returns data)}
}
\details{
Simply computes sensitivities based on cost function reduction (no permutation is performed ever).
         For plotting, this code uses ggplot graphics. If you wish to modify any part of the plot set mode="ggRet", the code
         will then return a ggplot structure that you can modify. Refer to ggplot documentation for more details.
}
\author{
Ogy Grujic (\email{ognjengr@gmail.com})
}
