% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftree.R
\name{ftree}
\alias{ftree}
\title{Display a Fancy File Tree}
\usage{
ftree(path = NULL, files = TRUE, unicode = TRUE)
}
\arguments{
\item{path}{character string specifying the file path to start the file tree. Defaults to NULL which displays the current working directoy.}

\item{files}{Logical. TRUE shows individual files, FALSE will only show folders. Default is TRUE.}

\item{unicode}{Show unicode emojis. Not sure how well this works across systems. Defaults to TRUE}
}
\value{
None
}
\description{
Outputs a formatted file tree structure in the console
}
