% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ftree}
\alias{ftree}
\title{Fits a regression tree to functional data}
\usage{
ftree(.X = NULL, .Y = NULL, .D = NULL, SIGMA_inv = NULL,
  cost.type = "sse", tree.type = "single", nP = if (tree.type ==
  "randomforest") round((ncol(.X)/3)) else ncol(.X), nBoot = 1000,
  .minSplit = 20, .minBucket = round(.minSplit/3), .cp = 0.005,
  verbose = TRUE, parallel = TRUE)
}
\arguments{
\item{.X}{- An nxp matrix of covariates (n= # of observations, p=# of covariates)}

\item{.Y}{- A matrix of functions stacked in columns. Assumes that all functions were evaluated on the same time grid}

\item{.D}{- Optional distance matrix for wss/rdist cost function}

\item{cost.type}{- Cost function type. It can be any of the following: "sse", "mahalanobis", "wss", "l2norm", "rdist" see details.}

\item{tree.type}{- What type of tree based predictor you want to fit. Currently supported: single tree, random forest, bagging}

\item{nP}{- The number of predictors to consider on each attempted split. Active only for tree.type = "randomforest"}

\item{nBoot}{- The number of trees to consider in bootstrapping}

\item{.minSplit}{- minimum required number of elements in a node in order to attempt a split.}

\item{.minBucket}{- minimum number of elements in leaf nodes. Defaults to .minSplit/3.}

\item{.cp}{- complexity parameter, split is accepted if it provides imporovement that is at least cp*rootGoodness}

\item{verbose}{- print progres (default = TRUE)}
}
\details{
This code implements various functional and multivariate tree splitting routines. See vignette for a detailed description of
         each cost function and for a tutorial on how to use this function.
}
\author{
Ognjen Grujic (\email{ogyg@stanford.edu} or \email{ognjengr@gmail.com})
}
