% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{matToList}
\alias{matToList}
\alias{matToList}
\title{A helper function for parallel computing (not exported)}
\usage{
matToList(.matrix, .chunksize)

matToList(.matrix, .chunksize)
}
\arguments{
\item{.matrix}{- A matrix to split into list;}

\item{.chunksize}{- A number of rows in each list element;}

\item{.matrix}{- A matrix you wish to divide into a list}

\item{.chunksize}{- The number of rows of \code{.matrix} to be contained in each
element of the resulting list.}
}
\details{
This code simply splits the rows of \code{.matrix} into a list where each element of the list
         contains \code{.chunksize} number of rows of the input matrix. If \code{.chunksize > nrow(.matrix)} the code
         will return a list with one element, the \code{.matrix} itself.

This function simply splits the provided matrix into \code{N} elements, each containing
         \code{.chunksize} of rows of the original matrix. The list elements are mutually exclusive.
}
\author{
Ognjen Grujic (\email{ognjengr@gmail.com})

Ogy Grujic (\email{ogyg@stanford.edu})
}
