% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plotFtree}
\alias{plotFtree}
\title{Plots a functional regression tree}
\usage{
plotFtree(.object, .index = 1, .horizontal = FALSE, .lineWidth = 1,
  .depth = Inf, .labSize = 1, .labAng = 0, .ggReturn = FALSE,
  .ylimLow = 0, .ylimHigh = NA, .node = NULL)
}
\arguments{
\item{.object}{- An fTree object or .segments object produced with \code{extractNodeData}}

\item{.index}{- Index of the tree you wish you plot}

\item{.horizontal}{- Whether to plot tree with horizontal orientation}

\item{.lineWidth}{- Width of the tree lines}

\item{.depth}{- Maximum plotting depth (useful for large trees)}

\item{.labSize}{- size of node labels}

\item{.labAng}{- Angle of node labels}

\item{.ggReturn}{- Whether to return an ggplot object or not.}
}
\details{
This function relies on "extractNodeData" function and ggplot plotting package
         output of this function are publication ready graphics of fitted regression trees.
}
\author{
Ognjen Grujic (\email{ogyg@stanford.edu})
}
